added yazi configs

+++ .local/share/scripts/chwp.sh
@@ -1,17 +1,18 @@
 #!/usr/bin/bash
 
-wallpapersDir=~/.local/share/wallpapers/
+WALLPAPERDIR=~/.local/share/wallpapers/
 
-magick "$1" "$wallpapersDir""current.jpg"
+cd $WALLPAPERDIR
+WALLPAPER="$(eza | fuzzel --dmenu)"
 
-wallust run -s "$wallpapersDir""current.jpg" >/dev/null 2>&1
+if [[ -z $WALLPAPER ]]; then
+  notify-send 'No wallpaper selected'
+  exit
+fi
+
+wallust run -s "$WALLPAPERDIR$WALLPAPER" >/dev/null 2>&1
 
 pidof waybar >/dev/null && killall waybar
 waybar &
 
-hyprctl hyprpaper unload all >/dev/null
-hyprctl hyprpaper preload ~/.local/share/wallpapers/current.jpg >/dev/null
-hyprctl hyprpaper wallpaper "DP-1,~/.local/share/wallpapers/current.jpg" >/dev/null
-hyprctl hyprpaper wallpaper "DP-2,~/.local/share/wallpapers/current.jpg" >/dev/null
-sleep 0.3
-hyprctl hyprpaper unload all >/dev/null
+swww img "$WALLPAPERDIR$WALLPAPER" --transition-type center
